+++ plugins/networkdelay/object.go
